pyproject.toml
[tool.poetry]
name = "talk-to-chart"
version = "0.1.0"
description = "AI-powered clinical chart query system for MDS coordinators"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "talk_to_chart", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
google-generativeai = "^0.3.0"
pydantic = "^2.5.0"
rich = "^13.7.0"
typer = "^0.9.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
talk-to-chart = "talk_to_chart.main:app"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true